+++ src/pages/Help/index.js
@@ -1,9 +1,11 @@
 import React from "react";
 import { View, Text } from "react-native";
+import Header from "../../pages/Header";
 
 export default function Help() {
     return (
         <View>
+            <Header />
             <Text>HELP</Text>
         </View>
     );