+++ src/pages/Login/index.js
@@ -4,19 +4,21 @@ import { View, Text, TouchableOpacity, TextInput, Image } from "react-native";
 import { Feather } from "@expo/vector-icons";
 
 import styles from "./styles";
-import arrow from "../../assets/arrow-right.png";
-import googleIcon from "../../assets/google.jpg";
-import faceIcon from "../../assets/face.png";
+import arrow from "../../assets/arrow.png";
+import emailIcon from "../../assets/Logar-Email.png";
+import googleIcon from "../../assets/logar-google.png";
+import faceIcon from "../../assets/logar-facebook.png";
+
+import Header from "../../pages/Header";
 
 export default function Login() {
     return (
         <View style={styles.loginContainer}>
+            <Header />
             <View style={styles.searchContainer}>
-                <Text style={styles.searchText}>
-                    Digite sua palavra chave:{" "}
-                </Text>
+                <Text style={styles.searchText}>Digite sua palavra chave:</Text>
                 <View style={styles.arrowText}>
-                    {/* <Image source={arrow} style={styles.arrow} /> */}
+                    <Image source={arrow} style={styles.arrow} />
                     <TextInput
                         style={styles.input}
                         placeholder="Ex: Web"
@@ -40,13 +42,13 @@ export default function Login() {
 
             <View style={styles.loginOpt}>
                 <TouchableOpacity onPress={() => console.log("mail clicked")}>
-                    <Feather name="mail" size={40} />
-                </TouchableOpacity>
-                <TouchableOpacity onPress={() => console.log("google clicked")}>
-                    <Image source={googleIcon} style={styles.googleIcon} />
+                    <Image source={emailIcon} style={styles.loginIcon} />
                 </TouchableOpacity>
                 <TouchableOpacity onPress={() => console.log("face clicked")}>
-                    <Image source={faceIcon} style={styles.faceIcon} />
+                    <Image source={faceIcon} style={styles.loginIcon} />
+                </TouchableOpacity>
+                <TouchableOpacity onPress={() => console.log("google clicked")}>
+                    <Image source={googleIcon} style={styles.loginIcon} />
                 </TouchableOpacity>
             </View>
         </View>